# ITE/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/materials/addDegree": {
            "post": {
                "tags": [
                    "Materials"
                ],
                "summary": "add degree for",
                "operationId": "20a172ed67f149ad95c5dbf22e84d75d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "material": {
                                            "type": "string"
                                        },
                                        "degree": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "example": [
                                {
                                    "material": "الفيزياء",
                                    "degree": 88
                                },
                                {
                                    "material": "التحليل 1",
                                    "degree": 75
                                }
                            ]
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "degree for material has been added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/materials/calcGBA": {
            "post": {
                "tags": [
                    "Materials"
                ],
                "summary": "calculate GBA for current user",
                "operationId": "606bb25e12ed3c281730fe0b777aab4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "academic_year",
                                    "year",
                                    "specialization"
                                ],
                                "properties": {
                                    "academic_year": {
                                        "type": "string",
                                        "example": "first_year"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "example": "common"
                                    },
                                    "year": {
                                        "type": "string",
                                        "example": "2022"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful calculated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "gba calculated seccessfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/materials/editDegree": {
            "put": {
                "tags": [
                    "Materials"
                ],
                "summary": "edit some degree for material",
                "operationId": "b9b1a5891f467b423c1636a6e59b0611",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_name",
                                    "material",
                                    "degree"
                                ],
                                "properties": {
                                    "user_name": {
                                        "type": "string",
                                        "example": "Harry Potter"
                                    },
                                    "material": {
                                        "type": "string",
                                        "example": "الفيزياء"
                                    },
                                    "degree": {
                                        "type": "integer",
                                        "example": "90"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully edit of degree",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "degree for material has been changed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/materials/getDegree/{material}": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "get degree for material",
                "operationId": "90455faf63924960179fa24dc4249263",
                "parameters": [
                    {
                        "name": "material",
                        "in": "path",
                        "description": "material name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "الفيزياء"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "degree": {
                                            "type": "integer",
                                            "example": "90"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/materials/getDegrees/{academic_year}/{specialization}": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "get degrees for specialization in some academic year",
                "operationId": "9a5fd14c0f6355f8da26723b7a15834f",
                "parameters": [
                    {
                        "name": "academic_year",
                        "in": "path",
                        "description": "academic year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "first_year"
                    },
                    {
                        "name": "specialization",
                        "in": "path",
                        "description": "specialization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "common"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/materials/getMaterials/{academic_year}/{specialization}": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "get materials for specialization in some academic year",
                "operationId": "35ba1540484df6649d178b3d6d043c7d",
                "parameters": [
                    {
                        "name": "academic_year",
                        "in": "path",
                        "description": "academic year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "first_year"
                    },
                    {
                        "name": "specialization",
                        "in": "path",
                        "description": "specialization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "common"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "in": "header",
                "name": "Authorization",
                "description": "Enter your Bearer token in the format `YOUR_TOKEN`. For example: `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`"
            }
        }
    }
}